API work, VoopAI work, busic economic, ubi, taxes, paying, some work on group roles permissions

+++ Database/Models/Districts/County.cs
@@ -17,7 +17,7 @@ public class County
 
     [VarChar(512)]
     public string? Description { get; set; }
-    public ulong Population { get; set;}
+    public int Population { get; set;}
 
     [EntityId]
     public string DistrictId { get; set;}
